Address Greptile review round 10: NWS city_timezone, discovery updated_at, snapshot savepoints, prediction Numeric
- Add city_timezone param to NwsClient.fetch_forecast for Liskov compatibility (P0)
- Add updated_at to discovery upsert set_ so status transitions are timestamped (P1)
- Wrap snapshot inserts in begin_nested() savepoints for per-item error isolation (P2)
- Change predicted_temp/std_dev from Float to Numeric(6,2) with migration (P1)

+++ shared/db/models.py
@@ -189,8 +189,8 @@ class Prediction(Base):
         nullable=False,
     )
     model_version: Mapped[str] = mapped_column(String(50), nullable=False)
-    predicted_temp: Mapped[float] = mapped_column(Float, nullable=False)
-    std_dev: Mapped[float] = mapped_column(Float, nullable=False)
+    predicted_temp: Mapped[Decimal] = mapped_column(Numeric(6, 2), nullable=False)
+    std_dev: Mapped[Decimal] = mapped_column(Numeric(6, 2), nullable=False)
     probability_distribution: Mapped[dict[str, Any] | None] = mapped_column(JSONB, nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), default=_utcnow, nullable=False)
